Split PSI aliases into another file
